refactor(mutation-kill): extract round-runner helper and split loop module (#1561, #1562)
run_for_file's 57-line inline body is split into a RunContext bundle plus a
_run_round helper (now 3 params, down from 9), and mutation_kill_loop.py's
insertion mechanics and headless-generation/CLI clusters move to new sibling
modules (mutation_kill_insert.py, mutation_kill_headless.py) — mirroring the
existing wrapper/report module split. Behavior is unchanged;
stryker_shard_pipeline.py now invokes mutation_kill_headless.py as the CLI
entry point.

A 17-agent review panel (run twice, after two rounds of fixes) surfaced and
resolved: a real mutual-import cycle from a leftover CLI shim in
mutation_kill_loop.py (removed), a cross-module boundary leak where the
commit-message builder reached into the insertion module's regex (now
returned via InsertOutcome.method_count), a monkeypatch-fragility risk from
an unqualified sibling-module import (now qualified, matching the existing
wrapper/mutation_report convention), and two naming inconsistencies
(prev_survivors -> prev_survivor_count, METHOD_RE -> TEST_METHOD_RE,
survivors -> survivor_count in _commit_message).

Several review findings were intentionally left unaddressed as pre-existing
or out of scope for this mechanical refactor: mutation_kill_loop_python.py's
own CC~11 loop body (same defect class, different file, no corresponding
issue), a handful of pre-existing test-coverage gaps predating the split, and
two architecture-level security observations (unattended commit-to-HEAD,
prompt-injection surface) that are properties of the whole mutation-kill
design, not of this split.

Closes #1561
Closes #1562
Part of #1568
